Added working query functionality

+++ DictionaryCodec/inc/Codec.h
@@ -20,13 +20,13 @@ public:
     bool EncodeColumnFile(const std::string& inputFile, const std::string& outputFile);
 
     // Query: Check if a data item exists in the encoded column, and return indices if it does
-    std::vector<unsigned int> QueryItem(const std::string& dataItem);
+    std::vector<size_t> QueryItem(const std::string& dataItem);
 
     // Query with Prefix: Search for items matching a prefix, returning unique items and their indices
     std::unordered_map<std::string, std::vector<size_t>> QueryByPrefix(const std::string& prefix) const;
 
     // Baseline Column Search (without dictionary encoding) for performance comparison
-    std::vector<unsigned int> BaselineSearch(const std::string& dataItem);
+    std::vector<size_t> BaselineSearch(const std::string& dataItem);
 
     // Helper to load encoded data from file
     void LoadEncodedFile(const std::string& inputFile);
@@ -40,11 +40,11 @@ public:
 
 private:
     // Dictionary and encoded data storage
-    std::unordered_map<std::string, int> dictionary_;               // Maps data items to unique integer codes
-    std::unordered_map<int, std::vector<unsigned int>> keyIndeces_; // Maps keys to indeces
-    std::unique_ptr<std::string[]> dataColumn_;                     // Unencoded data
-    std::vector<int> encodedColumn_;                                // Encoded column data as integers (keys)
-    mutable std::shared_mutex dictionaryMutex_;                     // Mutex for thread-safe access to dictionary
+    std::unordered_map<std::string, size_t> dictionary_;          // Maps data items to unique integer codes
+    std::unordered_map<size_t, std::vector<size_t>> keyIndeces_;  // Maps keys to indeces
+    std::unique_ptr<std::string[]> dataColumn_;                   // Unencoded data
+    std::vector<size_t> encodedColumn_;                           // Encoded column data as integers (keys)
+    mutable std::shared_mutex dictionaryMutex_;                   // Mutex for thread-safe access to dictionary
     size_t dataSize_;
 
     // Helper function to populate the dictionary using multiple threads
